cleaned up getAimAndMissPositions
